% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_query_params}
\alias{generate_query_params}
\title{Generate Query Parameters for specific API query}
\usage{
generate_query_params(query, ...)
}
\arguments{
\item{query}{string.  The specific API query desired.  Generally named the
same as associated functions but without underscores, so the request for
`getwebinardetails()` would be be "getwebinardetails".}

\item{...}{Named query parameter elements for desired api endpoint, such as
`next_page_token`.}
}
\value{
Query parameters to be passed to querying tools
}
\description{
This is based on the `generate_url()` function.
}
\keyword{internal}
