% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_url}
\alias{generate_url}
\title{Generate URL for specific API query by type and (if appropriate) ID}
\usage{
generate_url(query, ...)
}
\arguments{
\item{query}{string.  The specific API query desired.  Generally named the
same as associated functions but without underscores, so the request for
`getwebinardetails()` would be be "getwebinardetails".}

\item{...}{Named elements of URL for specific query desired, such as
`webinar_id`.}
}
\value{
Endpoint URL to be passed to querying tools
}
\description{
This is based on the function of the same name in the qualtRics package.
}
\keyword{internal}
