% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_access_token}
\alias{get_access_token}
\title{Get Access token}
\usage{
get_access_token(account_id, client_id, client_secret)
}
\arguments{
\item{account_id}{Zoom API app Account Id.}

\item{client_id}{Zoom API app Client Id.}

\item{client_secret}{Zoom API app Client Secret.}
}
\description{
Request a new token each time user executes a package function.
}
\keyword{internal}
