% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meeting_details.R
\name{get_meeting_details}
\alias{get_meeting_details}
\title{Get Meeting Details}
\usage{
get_meeting_details(meeting_id, account_id, client_id, client_secret)
}
\arguments{
\item{meeting_id}{Zoom Meeting ID (from list_meetings).}

\item{account_id}{Account Id granted by the Zoom developer app.}

\item{client_id}{Client Id granted by the Zoom developer app.}

\item{client_secret}{Client secret granted by the Zoom developer app.}
}
\value{
A data frame with the metadata about a meeting.
}
\description{
Get metadata about a single meeting including participant count, 
duration, and other meeting statistics.
}
\examples{
\dontrun{
dat <- get_meeting_details(meeting_id = "81753923023",
  your_account_id,
  your_client_id,
  your_client_secret)
}
}
\seealso{
See <https://marketplace.zoom.us/docs/api-reference/zoom-api/> for 
documentation on the Zoom API.
}
