% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zoom_api_request}
\alias{zoom_api_request}
\title{Send httr requests to Zoom API}
\usage{
zoom_api_request(
  verb = c("GET", "POST"),
  url = url,
  body = NULL,
  as = c("parsed", "raw"),
  token,
  query_params = query_params,
  ...
)
}
\arguments{
\item{verb}{Type of request to be sent (@seealso [httr::VERB()])}

\item{url}{Zoom endpoint URL created by [generate_url()] functions}

\item{...}{arguments passed to httr::content when parsing}
}
\description{
This is based on the function of the same name in the qualtRics package.
}
\keyword{internal}
