% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zoom_response_codes}
\alias{zoom_response_codes}
\title{Checks responses against Zoom response codes and returns error message.}
\usage{
zoom_response_codes(res)
}
\arguments{
\item{res}{results object from httr}
}
\description{
Checks responses against Zoom response codes and returns error message.
}
\keyword{internal}
