% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meeting_participants.R
\name{get_meeting_participants}
\alias{get_meeting_participants}
\title{Get Meeting Participants}
\usage{
get_meeting_participants(
  meeting_id,
  account_id,
  client_id,
  client_secret,
  page_size = 300
)
}
\arguments{
\item{meeting_id}{Zoom Meeting ID (from list_meetings). Note: Use meeting_id, not UUID.}

\item{account_id}{Account Id granted by the Zoom developer app.}

\item{client_id}{Client Id granted by the Zoom developer app.}

\item{client_secret}{Client secret granted by the Zoom developer app.}

\item{page_size}{Number of records per page. Default is 300.}
}
\value{
A data frame with data on each participant at a meeting.
}
\description{
Get participant info about a single meeting.
}
\examples{
\dontrun{
dat <- get_meeting_participants(meeting_id = "81753923023",
  your_account_id,
  your_client_id,
  your_client_secret)
}
}
\seealso{
See <https://marketplace.zoom.us/docs/api-reference/zoom-api/> for 
documentation on the Zoom API.
}
