% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_meetings.R
\name{list_meetings}
\alias{list_meetings}
\title{Get List of Meetings for a User}
\usage{
list_meetings(
  user_id,
  account_id,
  client_id,
  client_secret,
  meeting_type = "previous_meetings",
  page_size = 300
)
}
\arguments{
\item{user_id}{Zoom User Id.}

\item{account_id}{Account Id granted by the Zoom developer app.}

\item{client_id}{Client Id granted by the Zoom developer app.}

\item{client_secret}{Client secret granted by the Zoom developer app.}

\item{meeting_type}{Type of meetings to retrieve. Options are "scheduled", 
"live", "upcoming", or "previous_meetings". Default is "previous_meetings".}

\item{page_size}{Number of records per page. Default is 300.}
}
\value{
A data frame with all of the meetings hosted by a specific user.
}
\description{
Get list of meetings for a User. This is used to get the meeting Id's and
UUIDs to pass into other functions.
}
\examples{
\dontrun{
dat <- list_meetings(user_id = "user_id_string",
  your_account_id,
  your_client_id,
  your_client_secret,
  meeting_type = "previous_meetings")
}
}
\seealso{
See <https://marketplace.zoom.us/docs/api-reference/zoom-api/> for
documentation on the Zoom API.
}
