% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_meetings.R
\name{list_meetings}
\alias{list_meetings}
\title{List Zoom meetings}
\usage{
list_meetings(user_id)
}
\arguments{
\item{user_id}{the email or user id of your account}
}
\value{
a tibble
}
\description{
This functions returns a tibble with all the
zoom meetings in a account.
}
